Add errors for DELETE

+++ routes/task.delete.js
@@ -1,20 +1,33 @@
 const { Router } = require("express");
-const { Task } = require('../models')
+const { Task } = require('../models');
+const { param, validationResult } = require('express-validator');
+const { ErrorHandler } = require('../error')
+
 const router = Router();
 
 
 
-router.delete('/task/:id', async (req, res) => {
-    const id = req.params.id;
-    console.log('ID', typeof id);
-    try {
-        const user = await Task.destroy({where: 
-            { uuid: id }
+router.delete('/task/:uuid', 
+    param('uuid').isUUID(),
+    async (req, res, next) => {
+        try {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            throw new ErrorHandler(400, "invalid request", errors.array())
+        };
+
+        const uuid = req.params.uuid;
+
+        const taskID = await Task.findOne({where: {uuid: uuid}});
+        if(!taskID){
+            throw new ErrorHandler(422, "Can`t find task")
+        };
+        const task = await Task.destroy({where: 
+            { uuid: uuid }
         });
-        return res.json(user)
+        return res.json(task)
     } catch (err) {
-        console.log(err);
-        return res.send("Cant delete task")
+        next(err);
     };
 });
